+++ test/lib/el/el_test.dart
@@ -1,90 +1,124 @@
 import 'package:flutter_test/flutter_test.dart';
+import 'package:petitparser/debug.dart';
 
-import 'package:xwidget/src/el/parser.dart';
+import 'package:xwidget/xwidget.dart';
 
 
 main() {
-
-  final data = <String, dynamic>{
+    final data = <String, dynamic>{
     "size" : {
       "width": 300.0,
       "height": 200
     },
-    "indexes": [1, 0],
+    "indexes": [1, 0, 2],
     "providers": [
       { "title": "Title #1" },
       { "title": "Title #2" }
+    ],
+  };
+  final globalData = <String, dynamic>{
+    "users" : [
+      {
+        "abc": { "name": "Joe" }
+      },
+      {
+        "xyz": { "name": "Mike" }
+      },
+      {
+        "123": { "name": "Sally" }
+      }
     ]
   };
 
+  test('Assert global reference parsing', () {
+    final definition = ELParserDefinition(data: data, globalData: globalData);
+    final parser = definition.build(start: definition.expression);
+    final result = trace(parser).parse('global.users[indexes[1]].abc.name');
+    expect(result.value.evaluate(), "Joe");
+  });
+
+  test('Assert error global reference parsing', () {
+    final definition = ELParserDefinition(data: data, globalData: globalData);
+    final parser = definition.build(start: definition.expression);
+    final result = parser.parse('global.users[indexes[2]].123.name');
+    print("failure=${result.isFailure}");
+    expect(result.value.evaluate(), "Sally");
+  });
+
   test('Assert reference parsing', () {
-    final definition = ELParserDefinition(data: data);
+    final definition = ELParserDefinition(data: data, globalData: globalData);
     final parser = definition.build(start: definition.expression);
     expect(parser.parse('providers[indexes[0]].title').value.evaluate(), "Title #2");
   });
 
+  test('Assert reference parsing error', () {
+    final definition = ELParserDefinition(data: data, globalData: globalData);
+    final parser = definition.build(start: definition.expression);
+    expect(parser.parse('providers[indexes[0]].sku').value.evaluate(), "Title #2");
+  });
+
   test('Assert String reference concatenation', () {
-    final definition = ELParserDefinition(data: data);
+    final definition = ELParserDefinition(data: data, globalData: globalData);
     final parser = definition.build(start: definition.expression);
     expect(parser.parse("providers[indexes[0]].title + ', ' + providers[0].title").value.evaluate(), "Title #2, Title #1");
   });
 
   test('Assert expression in array reference', () {
-    final definition = ELParserDefinition(data: data);
+    final definition = ELParserDefinition(data: data, globalData: globalData);
     final parser = definition.build(start: definition.expression);
-    expect(parser.parse("providers[3 % length(indexes)].title").value.evaluate(), "Title #2");
+    expect(parser.parse("providers[4 % length(indexes)].title").value.evaluate(), "Title #2");
   });
 
   test('Assert addition/subtraction precedence', () {
-    final definition = ELParserDefinition(data: data);
+    final definition = ELParserDefinition(data: data, globalData: globalData);
     final parser = definition.build(start: definition.expression);
     expect(parser.parse('3 + 5 - 2').value.evaluate(), 6);
   });
 
   test('Assert addition/subtraction/multiplication precedence', () {
-    final definition = ELParserDefinition(data: data);
+    final definition = ELParserDefinition(data: data, globalData: globalData);
     final parser = definition.build(start: definition.expression);
     expect(parser.parse('3 + 5 - 2 * 2').value.evaluate(), 4);
   });
 
   test('Assert parentheses precedence', () {
-    final definition = ELParserDefinition(data: data);
+    final definition = ELParserDefinition(data: data, globalData: globalData);
     final parser = definition.build(start: definition.expression);
     expect(parser.parse('3 + (5 - 2) * 2').value.evaluate(), 9);
   });
 
   test('Assert num literal less than num literal', () {
-    final definition = ELParserDefinition(data: data);
+    final definition = ELParserDefinition(data: data, globalData: globalData);
     final parser = definition.build(start: definition.expression);
     expect(parser.parse('100 < 200').value.evaluate(), true);
   });
 
   test('Assert num literal greater than num literal', () {
-    final definition = ELParserDefinition(data: data);
+    final definition = ELParserDefinition(data: data, globalData: globalData);
     final parser = definition.build(start: definition.expression);
     expect(parser.parse('200.0 > 100').value.evaluate(), true);
   });
 
   test('Assert num reference compare to num literal', () {
-    final definition = ELParserDefinition(data: data);
+    final definition = ELParserDefinition(data: data, globalData: globalData);
     final parser = definition.build(start: definition.expression);
     expect(parser.parse('size.width > 200').value.evaluate(), true);
   });
 
   test('Assert int division', () {
-    final definition = ELParserDefinition(data: data);
+    final definition = ELParserDefinition(data: data, globalData: globalData);
     final parser = definition.build(start: definition.expression);
     expect(parser.parse('6 / 3').value.evaluate(), 2);
   });
 
   test('Assert int division', () {
-    final definition = ELParserDefinition(data: data);
+    final definition = ELParserDefinition(data: data, globalData: globalData);
     final parser = definition.build(start: definition.expression);
     expect(parser.parse('0 / 3').value.evaluate(), 0);
   });
 
   test('Assert int division by 0 exception', () {
-    final definition = ELParserDefinition(data: data);
+    final definition = ELParserDefinition(data: data, globalData: globalData);
     final parser = definition.build(start: definition.expression);
     expect(() => parser.parse('3 / 0').value.evaluate(), throwsException);
   });
@@ -98,7 +132,7 @@ main() {
       "isTrue" : true,
       "isFalse": false,
     };
-    final definition = ELParserDefinition(data: data);
+    final definition = ELParserDefinition(data: data, globalData: globalData);
     final parser = definition.build(start: definition.expression);
     expect(parser.parse('isTrue || isFalse && isFalse').value.evaluate(), true);
   });
@@ -112,7 +146,7 @@ main() {
       "item1": "String1",
       "item2": "String2",
     };
-    final definition = ELParserDefinition(data: data);
+    final definition = ELParserDefinition(data: data, globalData: globalData);
     final parser = definition.build(start: definition.expression);
     expect(parser.parse('item1 ?? item2').value.evaluate(), "String1");
   });
@@ -122,7 +156,7 @@ main() {
       "item1": null,
       "item2": "String2",
     };
-    final definition = ELParserDefinition(data: data);
+    final definition = ELParserDefinition(data: data, globalData: globalData);
     final parser = definition.build(start: definition.expression);
     expect(parser.parse('item1 ?? item2').value.evaluate(), "String2");
   });
@@ -133,7 +167,7 @@ main() {
       "item2": null,
       "item3": "String3"
     };
-    final definition = ELParserDefinition(data: data);
+    final definition = ELParserDefinition(data: data, globalData: globalData);
     final parser = definition.build(start: definition.expression);
     expect(parser.parse('item1 ?? item2 ?? item3').value.evaluate(), "String3");
   });